Added a separate test file for running via terminal
The existing test file works fine with PyCharm IDE by not via terminal as it depends on which path we run the unit test. Tried fixing them, setting appropriate path for choosing input files but did not help. The simplest fix for now is maintaining two test files, one if running via pycharm and the other if running via terminal